requirements.txt
appnope==0.1.2
backcall==0.2.0
bokeh==2.3.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
ddt==1.4.2
decorator==4.4.2
googledrivedownloader==0.4
idna==2.10
imageio==2.9.0
ipython==7.25.0
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
livelossplot==0.5.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
networkx==2.5.1
numpy==1.21.0
packaging==21.0
pandas==1.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-louvain==0.15
pytz==2021.1
PyYAML==5.4.1
rdflib==5.0.0
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
threadpoolctl==2.1.0
torch==1.9.0
torch-cluster==1.5.9
torch-geometric==1.7.2
torch-scatter==2.0.7
torch-sparse==0.6.10
torch-spline-conv==1.2.1
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
